+++ novelsave/client/bots/discord/containers.py
@@ -1,7 +1,19 @@
 from dependency_injector import containers, providers
 
+from novelsave.services.cloud.filehost import AnonFilesHost, GoFilesHost, NoneFilesHost
 from .endpoints import DownloadHandler, SearchHandler
 from .session import SessionHandler, Session
+
+
+class CloudContainer(containers.DeclarativeContainer):
+    config = providers.Configuration(strict=True)
+
+    filehost = providers.Selector(
+        config.cloud.filehost,
+        anonfiles=providers.Factory(AnonFilesHost),
+        gofiles=providers.Factory(GoFilesHost),
+        none=providers.Factory(NoneFilesHost),
+    )
 
 
 class SessionContainer(containers.DeclarativeContainer):
@@ -31,5 +43,10 @@ class DiscordApplication(containers.DeclarativeContainer):
 
     session = providers.Container(
         SessionContainer,
+        config=discord_config,
+    )
+
+    cloud = providers.Container(
+        CloudContainer,
         config=discord_config,
     )